refactor: change teacher view to adminview, etc
