+++ package.json
@@ -9,7 +9,7 @@
   "homepage": ".",
   "license": "MIT",
   "dependencies": {
-    "@graasp/apps-query-client": "github:graasp/graasp-apps-query-client.git",
+    "@graasp/apps-query-client": "github:graasp/graasp-apps-query-client.git#18/appSettings",
     "@material-ui/core": "4.12.3",
     "@material-ui/icons": "4.11.2",
     "@testing-library/jest-dom": "5.16.2",
