+++ src/components/App.js
@@ -1,35 +1,33 @@
 import React, { useContext } from 'react';
-import StudentView from './modes/student/StudentView';
-import { DEFAULT_PERMISSION } from '../config/settings';
-import TeacherMode from './modes/teacher/TeacherMode';
 import Header from './layout/Header';
 import { TokenProvider } from './context/TokenContext';
 import { Context } from './context/ContextContext';
+import AnalyzerView from './views/admin/AnalyzerView';
+import { CONTEXTS } from '../config/contexts';
+import BuilderView from './views/admin/BuilderView';
+import PlayerView from './views/read/PlayerView';
 
 export const App = () => {
   const context = useContext(Context);
 
   const renderContent = () => {
-    switch (context?.get('permission', DEFAULT_PERMISSION)) {
-      // show teacher view when in producer (educator) mode
-      case 'write':
-      case 'admin':
-        // case permission:
+    switch (context?.get('context')) {
+      case CONTEXTS.BUILDER:
         return (
           <>
             <Header />
-            <TeacherMode />
+            <BuilderView />
           </>
         );
+      case CONTEXTS.ANALYZER:
+        return <AnalyzerView />;
 
-      // by default go with the consumer (learner) mode
-      case 'read':
+      case CONTEXTS.PLAYER:
       default:
         return (
           <>
-            {(context?.get('permission')?.headerVisible ||
-              context?.get('standalone')) && <Header />}
-            <StudentView />
+            {context?.get('standalone') && <Header />}
+            <PlayerView />
           </>
         );
     }
