+++ src/components/layout/Header.js
@@ -8,7 +8,7 @@ import RefreshIcon from '@material-ui/icons/Refresh';
 import { useTranslation } from 'react-i18next';
 import { ReactComponent as Logo } from '../../resources/logo.svg';
 import { queryClient, HOOK_KEYS } from '../../config/queryClient';
-import DownloadCsvButton from '../modes/teacher/DownloadCsvButton';
+import DownloadCsvButton from '../views/admin/DownloadCsvButton';
 import { Context } from '../context/ContextContext';
 import { PERMISSION_LEVELS } from '../../config/settings';
 import {
