+++ src/components/views/admin/AnalyzerView.js
